requirements.txt
beautifulsoup4==4.10.0
PyYAML==5.4.1
requests==2.26.0